ra: migrating to kotlin
