+++ rider/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.pss9.rideraid">
+          package="com.pss9.rider">
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
     <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
@@ -13,8 +13,8 @@
             android:label="@string/app_name"
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
-            android:name="com.pss9.rideraid.RiderAidApplication">
-        <activity android:name="com.pss9.rideraid.ui.activity.SplashActivity">
+            android:name=".RiderAidApplication">
+        <activity android:name=".ui.activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <action android:name="android.intent.action.VIEW"/>
@@ -22,10 +22,10 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="com.pss9.rideraid.ui.activity.RiderActivity"/>
+        <activity android:name=".ui.activity.RiderActivity"/>
 
         <!-- Preferences -->
-        <activity android:name="com.pss9.rideraid.ui.preferences.MainPreferenceActivity"/>
+        <activity android:name=".ui.preferences.MainPreferenceActivity"/>
     </application>
 
 </manifest>
