+++ rider/src/main/java/com/pss9/rider/ant/EventHandler.kt
@@ -1,0 +1,9 @@
+package com.pss9.rider.ant
+
+/**
+ * Created on 2019-07-16.
+ */
+
+interface EventHandler {
+    fun onEvent(event: BikeEvent)
+}
