+++ rider/src/main/java/com/pss9/rider/model/Cadence.kt
@@ -1,0 +1,3 @@
+package com.pss9.rider.model
+
+class Cadence(private val timestamp: Long, val cadence: Long)
