+++ rider/src/main/java/com/pss9/rider/presenter/PositionPresenter.kt
@@ -1,0 +1,7 @@
+package com.pss9.rider.presenter
+
+interface PositionPresenter : PresenterActions {
+    interface View {
+        fun updateDistance(distance: Double)
+    }
+}
