+++ rider/src/main/java/com/pss9/rider/presenter/PresenterActions.kt
@@ -1,0 +1,9 @@
+package com.pss9.rider.presenter
+
+interface PresenterActions {
+    val isActive: Boolean
+
+    fun start()
+
+    fun stop()
+}
