+++ rider/src/main/java/com/pss9/rider/presenter/TelemetryPresenter.kt
@@ -1,0 +1,9 @@
+package com.pss9.rider.presenter
+
+interface TelemetryPresenter : PresenterActions {
+    interface View {
+        fun updateCadence(cad: Long)
+
+        fun updateSpeed(spd: Long)
+    }
+}
