+++ rider/src/main/java/com/pss9/rider/presenter/TimePresenter.kt
@@ -1,0 +1,7 @@
+package com.pss9.rider.presenter
+
+interface TimePresenter : PresenterActions {
+    interface View {
+        fun updateTime(second: Long)
+    }
+}
