--- rider/src/main/java/com/pss9/rideraid/ant/EventHandler.kt
@@ -1,9 +1,0 @@
-package com.pss9.rideraid.ant
-
-/**
- * Created on 2019-07-16.
- */
-
-interface EventHandler {
-    fun onEvent(event: BikeEvent)
-}
