--- rider/src/main/java/com/pss9/rideraid/model/Cadence.kt
@@ -1,3 +1,0 @@
-package com.pss9.rideraid.model
-
-class Cadence(private val timestamp: Long, val cadence: Long)
