--- rider/src/main/java/com/pss9/rideraid/model/Speed.kt
@@ -1,3 +1,0 @@
-package com.pss9.rideraid.model
-
-class Speed(private val timestamp: Long, val cadence: Long)
