--- rider/src/main/java/com/pss9/rideraid/presenter/PositionPresenter.kt
@@ -1,7 +1,0 @@
-package com.pss9.rideraid.presenter
-
-interface PositionPresenter : PresenterActions {
-    interface View {
-        fun updateDistance(distance: Double)
-    }
-}
