--- rider/src/main/java/com/pss9/rideraid/presenter/PresenterActions.kt
@@ -1,9 +1,0 @@
-package com.pss9.rideraid.presenter
-
-interface PresenterActions {
-    val isActive: Boolean
-
-    fun start()
-
-    fun stop()
-}
