--- rider/src/main/java/com/pss9/rideraid/presenter/TelemetryPresenter.kt
@@ -1,9 +1,0 @@
-package com.pss9.rideraid.presenter
-
-interface TelemetryPresenter : PresenterActions {
-    interface View {
-        fun updateCadence(cad: Long)
-
-        fun updateSpeed(spd: Long)
-    }
-}
