--- rider/src/main/java/com/pss9/rideraid/presenter/TimePresenter.kt
@@ -1,7 +1,0 @@
-package com.pss9.rideraid.presenter
-
-interface TimePresenter : PresenterActions {
-    interface View {
-        fun updateTime(second: Long)
-    }
-}
